+++ src/AngleSharp.Io/Network/CookieFieldNames.cs
@@ -1,0 +1,65 @@
+namespace AngleSharp.Io.Network
+{
+    using System;
+
+    /// <summary>
+    /// A collection of constants found in a cookie.
+    /// </summary>
+    public static class CookieFieldNames
+    {
+        /// <summary>
+        /// The Comment attribute name.
+        /// </summary>
+        public static readonly String Comment = "Comment";
+
+        /// <summary>
+        /// The CommentURL attribute name.
+        /// </summary>
+        public static readonly String CommentUrl = "CommentURL";
+
+        /// <summary>
+        /// The Discard attribute name.
+        /// </summary>
+        public static readonly String Discard = "Discard";
+
+        /// <summary>
+        /// The Domain attribute name.
+        /// </summary>
+        public static readonly String Domain = "Domain";
+
+        /// <summary>
+        /// The Expires attribute name.
+        /// </summary>
+        public static readonly String Expires = "Expires";
+
+        /// <summary>
+        /// The Max-Age attribute name.
+        /// </summary>
+        public static readonly String MaxAge = "Max-Age";
+
+        /// <summary>
+        /// The Path attribute name.
+        /// </summary>
+        public static readonly String Path = "Path";
+
+        /// <summary>
+        /// The Port attribute name.
+        /// </summary>
+        public static readonly String Port = "Port";
+
+        /// <summary>
+        /// The Secure attribute name.
+        /// </summary>
+        public static readonly String Secure = "Secure";
+
+        /// <summary>
+        /// The Version attribute name.
+        /// </summary>
+        public static readonly String Version = "Version";
+
+        /// <summary>
+        /// The HttpOnly attribute name.
+        /// </summary>
+        public static readonly String HttpOnly = "HttpOnly";
+    }
+}