+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 0.12.0
+
+Released on Thursday, May 2 2019.
+
+- Reference latest AngleSharp
+- Added `InputFile` class as a standard `IFile` implementation
+- Added `AppendFile` extensions for `IHtmlInputElement`
+
 # 0.10.1
 
 Released on Monday, January 7 2019.